refactor(auth): move token handler to auth package and improve encapsulation
