+++ internal/auth/basic.go
@@ -1,10 +1,15 @@
 package auth
 
 import (
+	"crypto/rsa"
 	"encoding/base64"
+	"encoding/json"
 	"errors"
 	"log/slog"
+	"net/http"
 	"strings"
+
+	"oauth2-task/internal/token"
 )
 
 var (
@@ -15,8 +20,8 @@ var (
 	ErrInvalidCredentials = errors.New("invalid username or password")
 )
 
-// UserPool represents a collection of users and their credentials
-type UserPool map[string]string
+// userPool represents a collection of users and their credentials
+type userPool map[string]string
 
 // BasicAuth represents basic authentication credentials
 type BasicAuth struct {
@@ -108,4 +113,71 @@ func (ba *BasicAuth) ParseBasicAuth(authHeader string) error {
 	ba.Password = credentials[1]
 
 	return nil
+}
+
+// TokenResponse represents the OAuth2 token response
+type TokenResponse struct {
+	AccessToken string `json:"access_token"`
+	TokenType   string `json:"token_type"`
+	ExpiresIn   int    `json:"expires_in"`
+}
+
+// tokenError represents an OAuth2 error response
+type tokenError struct {
+	Error            string `json:"error"`
+	ErrorDescription string `json:"error_description,omitempty"`
+}
+
+// HandleToken processes OAuth2 token requests
+func HandleToken(keyPair *rsa.PrivateKey, userPool map[string]string) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if r.Method != http.MethodPost {
+			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
+			slog.Error("Method not allowed", "method", r.Method, "status", http.StatusMethodNotAllowed)
+			return
+		}
+
+		// Create BasicAuth instance with user pool
+		basicAuth := NewBasicAuth(userPool)
+
+		// Validate Basic Auth
+		if err := basicAuth.ParseBasicAuth(r.Header.Get("Authorization")); err != nil {
+			w.Header().Set("Content-Type", "application/json")
+			w.WriteHeader(http.StatusUnauthorized)
+			errorResponse := GetErrorResponse(err)
+			json.NewEncoder(w).Encode(errorResponse)
+			slog.Error("Authentication failed", "error", err)
+			return
+		}
+
+		// Create token generator
+		generator := token.NewGenerator(keyPair)
+
+		// Generate a real JWT token
+		tokenString, err := generator.GenerateToken()
+		if err != nil {
+			w.Header().Set("Content-Type", "application/json")
+			w.WriteHeader(http.StatusInternalServerError)
+			err := json.NewEncoder(w).Encode(ErrorResponse{
+				Error:            "server_error",
+				ErrorDescription: "Failed to generate token",
+			})
+			if err != nil {
+				slog.Error("Failed to generate Error Response for token generation", "error", err)
+				return
+			}
+			slog.Error("Failed to generate token", "error", err)
+			return
+		}
+
+		// Return the token response
+		response := TokenResponse{
+			AccessToken: tokenString,
+			TokenType:   "Bearer",
+			ExpiresIn:   3600,
+		}
+
+		w.Header().Set("Content-Type", "application/json")
+		json.NewEncoder(w).Encode(response)
+	}
 }
