+++ main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"crypto/rsa"
-	"encoding/json"
 	"log/slog"
 	"net/http"
 	"os"
@@ -44,75 +43,12 @@ func setup() {
 	userPool = userpool.Default()
 }
 
-// TokenResponse represents the OAuth2 token response
-type TokenResponse struct {
-	AccessToken string `json:"access_token"`
-	TokenType   string `json:"token_type"`
-	ExpiresIn   int    `json:"expires_in"`
-}
-
-// TokenError represents an OAuth2 error response
-type TokenError struct {
-	Error            string `json:"error"`
-	ErrorDescription string `json:"error_description,omitempty"`
-}
-
-func tokenHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-		slog.Error("Method not allowed", "method", r.Method, "status", http.StatusMethodNotAllowed)
-		return
-	}
-
-	// Create BasicAuth instance with user pool
-	basicAuth := auth.NewBasicAuth(userPool)
-
-	// Validate Basic Auth
-	if err := basicAuth.ParseBasicAuth(r.Header.Get("Authorization")); err != nil {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusUnauthorized)
-		errorResponse := auth.GetErrorResponse(err)
-		json.NewEncoder(w).Encode(errorResponse)
-		slog.Error("Authentication failed", "error", err)
-		return
-	}
-
-	// Create token generator
-	generator := token.NewGenerator(keyPair)
-
-	// Generate a real JWT token
-	tokenString, err := generator.GenerateToken()
-	if err != nil {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusInternalServerError)
-		err := json.NewEncoder(w).Encode(auth.ErrorResponse{
-			Error:            "server_error",
-			ErrorDescription: "Failed to generate token",
-		})
-		if err != nil {
-			slog.Error("Failed to generate Error Response for token generation", "error", err)
-			return
-		}
-		slog.Error("Failed to generate token", "error", err)
-		return
-	}
-
-	// Return the token response
-	response := TokenResponse{
-		AccessToken: tokenString,
-		TokenType:   "Bearer",
-		ExpiresIn:   3600,
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(response)
-}
-
 func main() {
 	setup()
 	server := &http.Server{Addr: ":8080"}
 	slog.Info("Starting server", "port", 8080)
-	http.HandleFunc("/token", tokenHandler)
+	http.HandleFunc("/token", auth.HandleToken(keyPair, userPool))
+	http.HandleFunc("/.well-known/jwks.json", auth.HandleJWKS)
 	if err := server.ListenAndServe(); err != nil {
 		slog.Error("Failed to start server", "error", err)
 	}
