+++ pkg/keys/go.mod
@@ -1,7 +1,3 @@
 module github.com/wunderkind2k1/auth-server-task/pkg/keys
 
 go 1.21
-
-require (
-	github.com/golang-jwt/jwt/v5 v5.2.2
-)
